user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beanie>=1.30.0",
    "fastapi-users[beanie]>=14.0.1",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "yarl>=1.20.1",
]
